+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Updated documentation layout (#206)
 * Fixed root-level update in `piral-riot`, `piral-solid`, and `piral-hyperapp`
 * Fixed multi-asset output handling in `debug-pilet` with `piral-cli-webpack` (#258)
+* Fixed handling of spaces in `piral-cli` arguments (#259)
 * Fixed ignoring scoped pathed externals (#263)
 * Added basic support for mounting modules in `piral-ng`
 * Added ability to add new menu types in `piral-menu`