+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Piral Changelog
 
-## 0.11.8 (tbd)
+## 0.11.8 (July 9, 2020)
 
 * Updated documentation layout (#206)
 * Fixed root-level update in `piral-riot`, `piral-solid`, and `piral-hyperapp`